requirements.txt
python-dotenv
langchain-openai
langchain-elasticsearch
elasticsearch
google-generativeai
moviepy
